Testing
- Add simple test to the application
- Refactor the application to accommodate the changes
- Configure CircleCI test workflow
