+++ package.json
@@ -6,16 +6,20 @@
   "scripts": {
     "start:dev": "webpack-cli serve --mode development --hot --open 'Google Chrome'",
     "build": "webpack --mode production",
-    "start": "node server/index.js",
-    "lint": "eslint server"
+    "start": "node server/app.js",
+    "lint": "eslint server",
+    "test": "mocha test --exit"
   },
   "author": "Godfrey Tutu",
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "chai": "^4.3.4",
+    "chai-http": "^4.3.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "mocha": "^9.0.1",
     "node-sass": "^5.0.0",
     "pg": "^8.5.1",
     "react": "^17.0.1",
