+++ server/dbConfig.js
@@ -1,0 +1,7 @@
+const config = {
+  test: 'TEST_DATABASE_URL',
+  dev: 'DEV_DATABASE_URL',
+  prod: 'DATABASE_URL',
+};
+
+module.exports = config;
